Change the return type of choose, and add a more overflow-friendly variant.
--HG--
extra : convert_revision : 2179ca4ac612008fb22278477fb5aa77be84cbb0

+++ Statistics/Distribution/Binomial.hs
@@ -51,7 +51,7 @@ instance D.Mean BinomialDistribution where
 
 density :: BinomialDistribution -> Double -> Double
 density (BD n p) x =
-    fromIntegral (n `choose` floor x) * p ** x * (1-p) ** (fromIntegral n-x)
+    (n `choose` floor x) * p ** x * (1-p) ** (fromIntegral n-x)
 
 cumulative :: BinomialDistribution -> Double -> Double
 cumulative d =
